+++ tests/Unit/Drivers/Gd/FontProcessorTest.php
@@ -21,16 +21,16 @@ final class FontProcessorTest extends BaseTestCase
     public function testBoxSizeGdOne(): void
     {
         $processor = new FontProcessor();
-        $size = $processor->boxSize('test', new Font());
+        $size = $processor->boxSize('test', new Font()->setSize(1));
         $this->assertInstanceOf(SizeInterface::class, $size);
-        $this->assertEquals(16, $size->width());
+        $this->assertEquals(20, $size->width());
         $this->assertEquals(8, $size->height());
     }
 
     public function testBoxSizeGdTwo(): void
     {
         $processor = new FontProcessor();
-        $size = $processor->boxSize('test', new Font('2'));
+        $size = $processor->boxSize('test', new Font()->setSize(2));
         $this->assertInstanceOf(SizeInterface::class, $size);
         $this->assertEquals(24, $size->width());
         $this->assertEquals(14, $size->height());
@@ -39,7 +39,7 @@ final class FontProcessorTest extends BaseTestCase
     public function testBoxSizeGdThree(): void
     {
         $processor = new FontProcessor();
-        $size = $processor->boxSize('test', new Font('3'));
+        $size = $processor->boxSize('test', new Font()->setSize(3));
         $this->assertInstanceOf(SizeInterface::class, $size);
         $this->assertEquals(28, $size->width());
         $this->assertEquals(14, $size->height());
@@ -48,7 +48,7 @@ final class FontProcessorTest extends BaseTestCase
     public function testBoxSizeGdFour(): void
     {
         $processor = new FontProcessor();
-        $size = $processor->boxSize('test', new Font('4'));
+        $size = $processor->boxSize('test', new Font()->setSize(4));
         $this->assertInstanceOf(SizeInterface::class, $size);
         $this->assertEquals(32, $size->width());
         $this->assertEquals(16, $size->height());
@@ -57,7 +57,7 @@ final class FontProcessorTest extends BaseTestCase
     public function testBoxSizeGdFive(): void
     {
         $processor = new FontProcessor();
-        $size = $processor->boxSize('test', new Font('5'));
+        $size = $processor->boxSize('test', new Font()->setSize(5));
         $this->assertInstanceOf(SizeInterface::class, $size);
         $this->assertEquals(36, $size->width());
         $this->assertEquals(16, $size->height());